feat: Implement PersonnelServiceCrud

+++ src/Application/DTOs/Personnel/BonusDto.cs
@@ -1,5 +1,5 @@
 public class BonusDto
 {
-    public decimal Bonus { get; set; }
+    public double Bonus { get; set; }
     public string Description { get; set; } = string.Empty;
 }
